web/secure-mfa-backup/challenge/requirements.txt
appdirs==1.4.4
asyncio==3.4.3
blinker==1.6.2
Bootstrap-Flask==2.2.0
certifi==2023.5.7
click==8.1.3
Deprecated==1.2.13
dominate==2.8.0
Flask==2.3.2
Flask-Limiter==3.3.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
Pillow==9.5.0
pyee==8.2.2
Pygments==2.15.1
pyotp==2.8.0
pyppeteer==1.0.2
pyzbar==0.1.9
rich==13.3.5
SQLAlchemy==2.0.15
tqdm==4.65.0
typing_extensions==4.6.0
urllib3==1.26.16
visitor==0.1.3
websockets==10.4
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
